images:print: Add ability to print specific label from Dockerfile
This PR facilitates https://jira.coreos.com/browse/ART-455, to produce
a list of images that have a specific label in their Dockerfile.

When using images:print you may provide a new template parameter
"{label}" in your formatting string. When providing a value to the new
'--label/-l' option you can have the value of the given label
substituted in. If the label does not exist then an empty string is
substituted.

In the future it would be cool if this could provide a way to
substitute several different labels, but this PR is focused on MVP,
not gold plating.