rpmcfg.py: create two tags, build, push one
Tito release logic around some RPMs, particularly `openshift`, looks for
tags that look like "vX.Y.Z" while the existing tagging creates
"pkg-X.Y.Z" and pushes it to the source repo.

Add a vtag that is not pushed (there does not appear to be any reason to
do this, and the previous push is left only because we aren't sure what
might break if we stop doing it) to make tito happy.