runtime and schema: group option to ignore source branch fallbacks
Fallbacks are used to pick up release branches as soon as they are
created; however for the future we would like to create the branches in
advance and switch to using them only when we declare master has moved
on to the next release.

For the time when the branches are created but not to be used, this
gives doozer a group.yaml option to ignore the prematurely-created
"target" branches (e.g. "release-4.1") and just use specified fallbacks
(typically "master").  Conversely it can ignore the existence of
fallbacks if that ever becomes useful.