correct misleading date format info
I might actually rather go in the direction of using the YYYY-MM-DD
format for the date, and simply reformat it when handing off to
errata-tool. But I could also see an argument that it's good for elliott
to use a consistent format as errata does.

Either way the help and errors should not give users the wrong format.