Set correct permissions on rcm-guest for tarball sources
Per https://projects.engineering.redhat.com/browse/RCM-68294,
it would be more convenient for RCM if the file permissions
of tarball sources include read for other (o+r).

This PR uses `shutils.copyfile` rather than `shutils.copy2` to honor the umask for creating local tarballs,
and changes the instructions of sending to rcm-guest to explicitly set `r` for files and `rx` for directories.