Fix build-sync adding non-manifest lists in multiarch mode
There was a variable being set in a higher more-persistent scope than
intended during the filtering stage. Now we explicitly reset it at the
beginning of that stage. This avoids thinking a pullspec is a manifest
list when it's actually just a single image.

This also update some of the output to be multi-arch specific.