Add functional tests (#160)
Add functional test suite for Doozer. The test suite works by invoking `doozer` as a subprocess then checking for the exit code and output.

Currently the test suite doesn't mock any services, such as dist-git, Brew, etc, but interacts with real production services. This is not perfect but can significantly help our Python 3 migration.

Some test functions are not currently implemented but only a stub with `FIXME` marked. This is because those Doozer commands are either hard to test or they are already broken.

See `functional_tests/README.md` for more information and known issues.