ART-1536: scan-sources should not count disabled images as changed
This bug happens because `mode: disabled` and `mode: wip` only disable autoloading that image.yml when `--images` is not specified, but it will be loaded anyway if another image defines the disabled image as a dependent.

This PR https://github.com/openshift/ocp-build-data/pull/304 has removed `openshift-enterprise-template-service-broker-operator` from the dependent list of `template-service-broker`.

To prevent this issue from happening again, we can skip disabled images when scanning sources unless `--load-disabled` is specified.

Signed-off-by: Yuxiang Zhu <yuxzhu@redhat.com>