[ART-1981] distinguish between enabled and shipping repos
There are new check that verify that if all the rpms on the images can
be accounted for, and that there are no unused repositories declared to
be enabled.

The repositories that are enabled, are configured through the
`enabled_repos` key in the image definition. Setting a repository here
has two effects:
1. It will be written in `.oit/{un,}signed.repo`, one of which will end
up as `/etc/yum.repos.d/ubi.repo` in the resulting image.
2. A metadata file in distgit, `content_sets.yml`, which is consumed by
TPS tests to verify if all the RPMs are accounted for.

This poses an issue with dependencies for BUILDER images, as the
repository definition file is for all images used during the build
process, and content_set.yml only affects the shipping artifact.

This change allows the following configuration in `images/*.yml`:
```yaml
enabled_repos:
- rhel-server-rpms
- rhel-server-optional-rpms
non_shipping_repos:
- rhel-server-optional-rpms
```

The end result is:
- both repositories are enabled in `/etc/yum.repos.d/`
- `rhel-server-optional-rpms` is _not_ present in `content_set.yml`