bz search: Stop filtering on version
At the moment, for all the active release branches in ocp-build-data,
only openshift-3.11 uses `version` in `bugzilla.yml` to narrow down the
results of the appropriate bugs. The result set is not really narrow, as
all active versions from that time were listed here. The real filtering
of bugs happens when selecting on `target_release`.

The `openshift-4.Y` branches do not have `version` specified in
`bugzilla.yml`.

There are reasons to recycle the name `version`, as the generic value
`unspecified` is not accepted anymore.

This commit enables the recycling, as it stops using the `version`
filter in queries.