CLOUDWF-4852: enable specifying EUS pulp repos
https://projects.engineering.redhat.com/browse/CLOUDWF-4852

Traditionally for proper evaluation of RPM contents for grading, we have
specified content sets used, which are backed by pulp repos. In the case
of EUS releases, we need to further specify the pulp repo, as there may
be multiple per content set.

This is implemented by allowing specifying in content_sets.yaml either
the name of the content set or of the repo. Both will be accepted in
OSBS; for EUS releases we will specify the repos. Either way doozer
still needs to validate that a content set or pulp repo by that name
exists before passing it through.