Make retries image builds configurable
In the past, we had nightly full product builds. Then it made sense to
retry a failed image build.

Currently, our automation detects that a build is needed, and will spin
off a new build when necessary.

In the case of an image that takes a lot of timne to build, say
installer-artifacts, it takes around 1.5 hours for a single build. This
means that with three failing retries, product builds are halted for
three or four hours.

With this change, `doozer images:build` will grow a command line option
`--build-retries` that defaults to `1`, so that, in the default case:
- retries will continue to happen through other means
- images that need rebuilding will get rebuilt faster

For builds on a schedule, like OCP3.11, it makes sense to set retries to
3.