promote: Print digests of arch-specific heterogeneous payloads
`artcd promote` print a JSON message through stdout for Jenkinsfile
to pass the information of the promoted release payloads.

This PR adds more fields to the JSON message to include digests of arch-specific heterogeneous payloads
so that the Jenkinsfile will be able to sign heterogeneous payloads
using those digests without the need to pull the payloads again. Pulling
the payloads before signing could be a vulnerability.