[ART-4558] Use repoquery to detect non-latest rpms
Currently we detect outdated rpms in images and RHCOS by
comparing the NVRs of the installed rpms to the NVRs of the rpms in our
`rhaos-x.y-rhel-z-candidate` Brew tag.

The current approach has 2 issues. See ART-4558 and ART-3965.

With this new approach, `repoquery` command will be used to query
available RPMs in the configured build repos instead of the candidate
Brew tag. This will provide more accurate detection and also be able to
detect rpms that are not in rhaos repos (e.g. RHEL repos).

Note this PR only covers the `gen-payload` change. A change for
`scan-sources` will be another PR.