mass-rebuilds: only for streams.<stream>.image attribute
If one introduces or changes an attribute in the stream definition in
streams.yml, this will result in a new config_digest for all consuming
images, which results in mass rebuilds. In the context of product
rebuilds, we are only interested if the key `image` of the streams entry
changes. This commit makes that happen.