OKD must be based on CentOS Stream and not contain ELS RHEL content. As such, images that ART builds in brew must be rebuilt using different, non-RHEL, base images.
To achieve this, ART is not responsible for building the
images directly, but instead generates Test Platform
ci-operator configuration files which, when a PR merges,
causes non-RHEL OKD image variants to be built.

ART generates these files as they are the source
of truth for which images compose the OCP release payload
(and therefore knows which images should be build &
included in OKD).

A new set of verbs under "images:okd" is responsible
for generating this ci-operator configuration based on
ART's OCP metadata. These OKD ci-operator configurations
will trigger prowjobs to be run after there is a merge in
repos that contribute to OCP (and thus OKD). These, so
called, postsubmit jobs, run after a merge, are used for
OCP as well to promote images to the ocp 4.x imagestreams.
The ci-operator configurations generated by the new verb
will be "variants" of the OCP postsubmit jobs - running
alongside the core OCP postsubmit jobs, and capable
of introducing small changes (like base image replacements)
into images that are ultimately promoted for OKD into the
origin namespace and scos-4.x imagestream.

Some of ART's ocp-build-data metadata needs
tweaks in order to inform OKD build instructions. For example,
"openshift-base-rhel9.yml" would normally supply a RHEL9
base image for ART built OCP images. However, when building
for OKD, we want ci-operator to actually base child images
on a CentOS Stream base image. This is achieved using the
new "okd_alignment" stanza in image metadata.

In the case of "openshift-base-rhel9.yml", a stanza will be
added like:
content:
  source:
    okd_alignment:
      resolve_as:
        stream: centos_stream9

Which will instruct the new doozer images:okd verb to base
child images of openshift-base-rhel9 on the image referred
to by a stream.yml entry called "centos_stream9".

okd_alignment also supports overriding ART metadata configuration
when it generates ci-operator configuration files for OKD builds.
- "dockerfile" in case it is impossible to create
  a single Dockerfile for building both OCP and OKD.
- "context_dir" in case ci-operator configuration needs a
  different context dir.
- "path" to the dockerfile.
- "build_args" in case special docker build arguments need to be passed
  into to ci-operator builds for the OKD variant.
- "inject_rpm_repositories" which can cause CI specific RPM repositories
  to be baked into images as they are built in CI.

streams.yml will also support okd specific tweaks. For example, if
we ever want to use CentOS based builder images instead of RHEL
based builder images, we can use "okd.resolve_as.image" in a
streams.yml entry. For example,
rhel-9-golang-1.20:
  image: openshift/golang-builder:v1.20.12-202403212140.el9.g2983c24.el9
  mirror: true
  ...
  okd:
    resolve_as:
      image: registry.ci.openshift.org/ocp/builder:centos-stream-9-golang-1.20-openshift-{MAJOR}.{MINOR}

This will instruct the stream to resolve for ci-operator configuration
based builds as the literal pullspec provided. No attempt will be made
to produce this image - doozer will simply assume it exists when
creating ci-operator configuration files - so another pipeline may
need to put the image there.

Eventually "images:okd prs open" will be run to continuously reconcile
with ci-operator configuration for OKD variants  in openshift/release.
But for now, it will simply make recommended changes atop a local clone of the
repository. It remains for a human to open a PR to merge the recommended
changes to github.

Another verb "images:okd check" is available to detect whether all
images in the app.ci scos-4.x imagestream (in the origin namespace) are
based on CentOS stream. A CSV is output reporting whether the images
seem to be compliant or if they contain RHEL content.