WIP: sweep earlier
This change (naively, wrongly) wants to move the moment of setting bugs
to QE right before the image rebase starts.

In the case of a long-running ocp4 job (let's say: Mass rebuild), the
time delta of rebase and when we look for bugs that are on MODIFIED,
becomes rather large. This means that bugs are set to ON_QA before the
commit is built.

This change proposes to do the bug transition before rebasing to
distgit.

This would not be necessary if https://issues.redhat.com/browse/ART-6566
were implemented.