prepare-release: Explicitly add newly created advisory to batch
Due to an issue in Errata Tool, if batch is enabled for an ET Release,
newly created advisories will be implicitly added to an open batch.

This has caused chaos because our ET release (RHOSE ASYNC - AUTO) is
shared among multiple teams (ART, CPaaS, etc) and advisories for
different OCP releases are prepared in parallel.

I would like to propose a new CLI option `--batch-id` for the `elliott
create` command, so that we can explicitly specify the batch ID when
creating an advisory. To achieve this, raw Errata REST API is used since
the errata_tool library doesn't expose that field.

Also updated prepare_release job to use the new CLI option.