Fix a bug in image embargo detection
In function `_mapped_image_from_member`, the embargo state (private_fix
flag) is set to its parent member's state. This is wrong.

The assumption is that when one of the parent images has an embargo, the
child image is considered embargoed as well.

Also moving private_fix and image_build_method fields from ImageDistGitRepo class to Metadata.