Elliott: A new approach to check for shipped rpms
Currently, we consider an rpm shipped if it has `*-released` Brew tag.
Released rpms will be excluded from `find-builds` so they will not be
attached to Errata.

This has caused a few issues: Some external rpms (e.g. python-zipp) are
tagged into both rhaos-4.17-ironic-rhel-9-candidate and rhaos-4.16-ironic-rhel-9-candidate.
Once the rpm is shipped to 4.17, we will not be able to ship it to 4.16.

With the new approach, we read all candidate tags for an OCP minor
version from ocp-build-data erratatool.yml, they consider a rpm is
shipped if it has a tag without the -candidate suffix.

The feature is not enabled by default unless group config option
`feature_gates.product_version_aware_rpm_sweep` is enabled.