Fix env var propagation with telemetry enabled
When OpenTelemetry is enabled, environment variables `TRACEPARENT` needs
to be set for child process for trace context propagation.

There is a bug in `cmd_gather_async`: If parameter `env` is not set or
set to `None`, all environment variables of the current process will be
passed to the child process. However, when injecting `TRACEPARENT`, we
pass `{'TRACEPARENT': 'something'}` into child process without other
environment variables of the current process, causing misbehavior in the
child process.

This PR will fix this by copying all environment variables of the
current process if `env` is not set.