[Konflux] Fix missing SAST tasks errors for OLM bundles
The default value of `sast` parameter of
`KonfluxClient.start_pipeline_run_for_image_build` is `False`, which is
to remove all SAST tasks from the pipeline template. This leads to
confusion when `sast` parameter is not set. We forget to specify this
parameter for bundle and FBC build and saw unexpected EC violations.

I would like to propose such a fix:

1. The default value of `sast` is set to `None`, which means not
   changing the default behavior of the pipeline template.
2. When `sast` is set to True, assert the SAST tasks must exist in the
   pipeline template.
3. When `sast` is set to False (instead of None), remove all SAST tasks.