[Konflux] ART-13198: Snapshot creation for bundle and fbc builds
This PR makes it possible to create Konflux Snapshot(s) for a list of
mixed container images, OLM bundles, and FBC fragments using `elliott
snapshot new` command.

Multiple snapshot instances will be created if passed builds are from
multiple Konflux applications, one for each application.

This PR also removes `--for-fbc` option because we can determine whether
a build is FBC or not according to its NVR.

Note that this PR changes the way of getting Konflux application and
component for a build record. We should get those info from DB instead
of construct them because:

1. We might change our naming conventions.
2. We use different naming formats for regular images, bundles, and
   FBCs.