ART-14003: [Konflux] Ship RHCOS through Errata
Shipping RHCOS through Konflux is currently not supported. We need to
continue to ship RHCOS through Errata until RHCOS is fully migrated to
Konflux.

With this PR, `gen-assembly` will add a placeholder advisory `rhcos: -1` for RHCOS, and `prepare-release-konflux` will populate the RHCOS advisory when needed.

Note that currently permissive mode of `elliott find-bugs` is used to
sweep bugs to advisories. This is not perfect but this PR doesn't change
that.