[Konflux] fbc:merge: Search for all stage FBC fragments
When no input FBC fragments are provided, search for all related tags
(starting with `ocp__{MAJOR}.{MINOR}__*`) in the stage as input.

This simplifies the usage of this command, and will make sure all
operators will be included in the resulting index even if we don't have
a rebuild in a z-stream release.