Enable a new pipeline to run build-sync in multi-model mode The multi-arch teams want to be able to validate nightlies which are EC/RC candidates BEFORE they are promoted. These candidates are selected based on amd64 nightlies.
However, the release controllers are independent and create payloads
in an unsynchronized way, so there is no multi-arch payload that
completely matches the amd64 payload.

This approach will provide a compromise where multi-arch
payloads can be formulated to include the exact same source
commits / digests as the amd64 nightlies with a manifest-listed
multi-arch payload. This process is achieved by providing
gen-payload a "model" on which to build the multi-arch payload.

To keep the multi-arch payloads in sync with amd64, we will
need to run this periodically with --multi-model=amd64:0 which
will construct a multi nightly based on the most recent
amd64 nightly. If a multi nightly already exists with the
same name as the amd64 nightly, the task will exit early. In
other words, you should be able to run the task frequently
without creating too many multi payloads.

Once this new method is running regularly, the normal
build-sync can stop building and applying multi payload
changes.